+++ src/components/DraggableCard.tsx
@@ -1,29 +1,12 @@
-import { memo, useState, useEffect, CSSProperties, FC } from 'react';
+import { memo, useEffect, CSSProperties, FC } from 'react';
 import type { DragSourceMonitor } from 'react-dnd';
 import { useDrag } from 'react-dnd';
 import { getEmptyImage } from 'react-dnd-html5-backend';
-import { styled } from '@mui/material/styles';
-import Card from '@mui/material/Card';
-import CardHeader from '@mui/material/CardHeader';
-import CardContent from '@mui/material/CardContent';
-import Collapse from '@mui/material/Collapse';
-import Avatar from '@mui/material/Avatar';
-import IconButton, { IconButtonProps } from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { Theme } from '@mui/material/styles';
 import { ItemTypes } from '../constants';
+import { BasicCardProps } from '../interfaces';
+import BasicCard from './cards/BasicCard';
 
-interface ExpandButtonProps extends IconButtonProps {
-  expand: boolean;
-}
-
-interface BasicCard {
-  title: string;
-  preview?: boolean;
-}
-
-interface DraggableCardProps extends BasicCard {
+interface DraggableCardProps extends BasicCardProps {
   id: string;
   left: number;
   top: number;
@@ -47,46 +30,6 @@ function getStyles(left: number, top: number, isDragging: boolean): CSSPropertie
     height: isDragging ? 0 : '',
   };
 }
-
-const ExpandButton = styled((props: ExpandButtonProps) => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { expand, ...rest } = props;
-  return <IconButton {...rest} />;
-})(({ theme, expand }: { theme: Theme; expand: boolean }) => ({
-  transform: !expand ? 'rotate(0deg)' : 'rotate(180deg)',
-  marginLeft: 'auto',
-  transition: theme.transitions.create('transform', {
-    duration: theme.transitions.duration.shortest,
-  }),
-})) as React.ComponentType<ExpandButtonProps>;
-
-export const BasicCard = ({ title, preview }: BasicCard) => {
-  const [expanded, setExpanded] = useState(false);
-  const avatar = title.charAt(0).toUpperCase();
-
-  const toggleExpand = () => {
-    setExpanded(!expanded);
-  };
-
-  return (
-    <Card sx={{ maxWidth: 345 }} role={preview ? 'BoxPreview' : 'Box'}>
-      <CardHeader
-        avatar={<Avatar aria-label="avatar">{avatar}</Avatar>}
-        action={
-          <ExpandButton expand={expanded} onClick={toggleExpand} aria-expanded={expanded} aria-label="show more">
-            <ExpandMoreIcon />
-          </ExpandButton>
-        }
-        title={title}
-      />
-      <Collapse in={expanded} timeout="auto" unmountOnExit>
-        <CardContent>
-          <Typography paragraph>This is some test content.</Typography>
-        </CardContent>
-      </Collapse>
-    </Card>
-  );
-};
 
 const DraggableCard: FC<DraggableCardProps> = memo(function DraggableCard({
   id,
@@ -121,7 +64,7 @@ const previewStyles: CSSProperties = {
   opacity: 0.5,
 };
 
-export const CardDragPreview: FC<BoxDragPreviewProps> = memo(function CardDragPreview({ title }: BasicCard) {
+export const CardDragPreview: FC<BoxDragPreviewProps> = memo(function CardDragPreview({ title }: BasicCardProps) {
   return (
     <div style={previewStyles}>
       <BasicCard title={title} preview />
